Fix incorrectly parsed face numbers in accountplayers message.

+++ jxclient/trunk/src/jxclient/com/realtime/crossfire/jxclient/main/JXCWindow.java
@@ -351,7 +351,7 @@ public class JXCWindow extends JFrame {
        * {@inheritDoc}
        */
       @Override
-      public void addAccount(@NotNull final String name, @NotNull final String characterClass, @NotNull final String race, @NotNull final String face, @NotNull final String party, @NotNull final String map, final short level, final short faceNumber) {
+      public void addAccount(@NotNull final String name, @NotNull final String characterClass, @NotNull final String race, @NotNull final String face, @NotNull final String party, @NotNull final String map, final int level, final int faceNumber) {
           final CharacterInformation information = new CharacterInformation();
           information.setName(name);
           information.setRace(race);
